Move and export ActorRef and ActorRefFrom types
