+++ packages/core/actor/package.json
@@ -1,0 +1,7 @@
+{
+  "main": "dist/xstate.cjs.js",
+  "module": "dist/xstate.esm.js",
+  "preconstruct": {
+    "source": "../src/Actor"
+  }
+}
